Reorder code
